+++ BigImageViewer/src/main/java/com/github/piasy/biv/view/ImageViewFactory.java
@@ -38,10 +38,14 @@ public class ImageViewFactory {
 
     public final View createMainView(Context context, int imageType, File imageFile,
             int initScaleType) {
-        if (imageType == ImageInfoExtractor.TYPE_STILL_IMAGE) {
-            return createStillImageView(context);
-        } else {
-            return createAnimatedImageView(context, imageType, imageFile, initScaleType);
+        switch (imageType) {
+            case ImageInfoExtractor.TYPE_GIF:
+            case ImageInfoExtractor.TYPE_ANIMATED_WEBP:
+                return createAnimatedImageView(context, imageType, imageFile, initScaleType);
+            case ImageInfoExtractor.TYPE_STILL_WEBP:
+            case ImageInfoExtractor.TYPE_STILL_IMAGE:
+            default:
+                return createStillImageView(context);
         }
     }